Added migrations

+++ internal/database/order_item.go
@@ -7,11 +7,11 @@ type OrderItemModel struct {
 }
 
 type OrderItem struct {
-	ID        uint `gorm:"primaryKey;AUTO_INCREMENT"`
-	OrderID   uint `gorm:"not null"`
-	ProductID uint `gorm:"not null"`
-	Quantity  int  `gorm:"not null"`
-	Price     int  `gorm:"not null"`
+	ID        uint    `gorm:"primaryKey;AUTO_INCREMENT"`
+	OrderID   uint    `gorm:"not null"`
+	ProductID uint    `gorm:"not null"`
+	Quantity  int     `gorm:"not null"`
+	Price     float64 `gorm:"not null"`
 
 	Order   Order   `gorm:"foreignKey:OrderID;constraint:OnDelete:CASCADE"`
 	Product Product `gorm:"foreignKey:ProductID"`
